fix sonarquce duplicate code error
