+++ src/components/BMDashboard/WeeklyProjectSummary/WeeklyProjectSummary.jsx
@@ -179,6 +179,79 @@ function WeeklyProjectSummary() {
     }
   }, [dispatch, materials.length]);
 
+  const selectStyles = useMemo(
+    () => ({
+      control: base => ({
+        ...base,
+        backgroundColor: darkMode ? '#22272e' : '#fff',
+        borderColor: darkMode ? '#375071' : '#ccc',
+        color: darkMode ? '#fff' : '#232323',
+        minHeight: 38,
+        boxShadow: 'none',
+        borderRadius: 8,
+      }),
+      menu: base => ({
+        ...base,
+        backgroundColor: darkMode ? '#22272e' : '#fff',
+        fontSize: 12,
+        zIndex: 10001,
+        borderRadius: 8,
+        marginTop: 2,
+        color: darkMode ? '#fff' : '#232323',
+      }),
+      menuList: base => ({
+        ...base,
+        maxHeight: 400,
+        overflowY: 'auto',
+        backgroundColor: darkMode ? '#22272e' : '#fff',
+        color: darkMode ? '#fff' : '#232323',
+        padding: 0,
+      }),
+      option: (base, state) => ({
+        ...base,
+        backgroundColor: state.isSelected
+          ? '#0d55b3'
+          : state.isFocused
+          ? '#0d55b3'
+          : darkMode
+          ? '#22272e'
+          : '#fff',
+        color: state.isSelected ? '#fff' : darkMode ? '#fff' : '#232323',
+        fontSize: 13,
+        padding: '10px 16px',
+        cursor: 'pointer',
+      }),
+      singleValue: base => ({
+        ...base,
+        color: darkMode ? '#fff' : '#232323',
+      }),
+      multiValue: base => ({
+        ...base,
+        backgroundColor: darkMode ? '#375071' : '#e2e7ee',
+        borderRadius: 6,
+        fontSize: 12,
+        marginRight: 4,
+      }),
+      multiValueLabel: base => ({
+        ...base,
+        color: darkMode ? '#fff' : '#333',
+        fontSize: 12,
+        padding: '2px 6px',
+      }),
+      multiValueRemove: base => ({
+        ...base,
+        color: darkMode ? '#fff' : '#333',
+        ':hover': {
+          backgroundColor: darkMode ? '#0d55b3' : '#e2e7ee',
+          color: '#fff',
+        },
+        borderRadius: 4,
+        padding: 2,
+      }),
+    }),
+    [darkMode],
+  );
+
   const quantityOfMaterialsUsedData = useMemo(() => {
     if (!materials.length) return [];
     const uniqueMaterials = Array.from(new Map(materials.map(m => [m._id, m])).values());
@@ -198,7 +271,7 @@ function WeeklyProjectSummary() {
         title: 'Risk profile for projects',
         key: 'Risk profile for projects',
         className: 'full',
-        content: <ProjectRiskProfileOverview />,
+        content: <ProjectRiskProfileOverview selectStyles={selectStyles} />,
       },
       {
         title: 'Project Status',
@@ -251,9 +324,14 @@ function WeeklyProjectSummary() {
         content: [1, 2, 3].map((_, index) => {
           let content;
           if (index === 1) {
-            content = <QuantityOfMaterialsUsed data={quantityOfMaterialsUsedData} />;
+            content = (
+              <QuantityOfMaterialsUsed
+                data={quantityOfMaterialsUsedData}
+                selectStyles={selectStyles}
+              />
+            );
           } else if (index === 2) {
-            content = <TotalMaterialCostPerProject />;
+            content = <TotalMaterialCostPerProject selectStyles={selectStyles} />;
           } else {
             content = <p>📊 Card</p>;
           }
